# docs/swagger.yaml
basePath: /api/v1
definitions:
  apierror.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  apierror.SuccessResponse:
    properties:
      status:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /transactions/export-csv:
    get:
      description: Exports transactions based on filters as a CSV file
      parameters:
      - description: Transaction ID
        in: query
        name: transaction_id
        type: integer
      - description: Terminal ID
        in: query
        name: terminal_id
        type: string
      - description: Payment Narrative
        in: query
        name: payment_narrative
        type: string
      - description: Start Date
        in: query
        name: date_from
        type: string
      - description: End Date
        in: query
        name: date_to
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: File transfer
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      summary: Export transactions as CSV
      tags:
      - transactions
  /transactions/export-json:
    get:
      description: Exports transactions based on filters as a JSON file
      parameters:
      - description: Transaction ID
        in: query
        name: transaction_id
        type: integer
      - description: Terminal ID
        in: query
        name: terminal_id
        type: string
      - description: Payment Narrative
        in: query
        name: payment_narrative
        type: string
      - description: Start Date
        in: query
        name: date_from
        type: string
      - description: End Date
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: File transfer
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      summary: Export transactions as JSON
      tags:
      - transactions
  /transactions/upload:
    post:
      consumes:
      - multipart/form-data
      description: Uploads a CSV file containing transactions
      parameters:
      - description: CSV file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apierror.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      summary: Upload CSV file
      tags:
      - transactions
swagger: "2.0"